READY
HA HA HA

+++ game.h
@@ -15,7 +15,7 @@ class Game : public QObject
 
 private:
     Base base;
-    int points;
+    unsigned long long points;
     pair <QChar, bool> alphabet[32];
     QString playerName;
 
@@ -33,8 +33,8 @@ public slots:
     void resetAlphabet();
     void resetGame();
     void setName(QString name);
-    void setPoints(int points);
-    int getPoints();
+    void setPoints(unsigned long long points);
+    unsigned long long getPoints();
     QString getLetter(int index);
     QString getAnswer();
     QString getQuestion();
